+++ changes.md
@@ -11,6 +11,7 @@
 8. 3-dimensional TTree::Draw now produces TPolyMarker3D by default
 9. Fix - multipart message handling (#250)
 10. Fix - interactive move/resize on touch devices
+11. Provide `settings.PreferSavedPoints` to exclude function evaluation when there are saved points
 
 
 ## Changes in 7.2.0